requirements.txt
faiss-cpu
numpy
requests
tqdm
datasets